Current dump of papers
In a weird json format, I parsed myself the Openreview data. I intent to
use it as a source/database to generate all the subsequent paper list
and paper dedicated pages (coming soon, for the virtual MIDL).

The rationale to commit that one here is that we will probably have to
edit it by hand, for some corner cases. Hence, we won't be able to fully
generate it from some source